move database features from __init__ to database.py
