+++ flask_app/__init__.py
@@ -1,37 +1,14 @@
 """
 The flask application package.
 """
-import sqlite3
-from contextlib import closing
 
-from flask import Flask, g
+from flask import Flask
+
 DATABASE = 'C:\\Users\\Andreas\\dev\\kidsakoder-minecraft\\flask_app\\tmp\\database.db'
 DEBUG = True
 app = Flask(__name__)
 app.config.from_object(__name__)
 
 
-def connect_db():
-    return sqlite3.connect(app.config['DATABASE'])
-
-
-def init_db():
-    with closing(connect_db()) as db:
-        with app.open_resource('schema.sql', mode='r') as f:
-            db.cursor().executescript(f.read())
-        db.commit()
-
-
-@app.before_request
-def before_request():
-    g.db = connect_db()
-
-
-@app.teardown_request
-def teardown_request(exception):
-    db = getattr(g, 'db', None)
-    if db is not None:
-        db.close()
-
-
+import flask_app.database
 import flask_app.views
